[#3418] move discovery outside of views
