+++ hs_rest_api/discovery.py
